requirements.txt
aiofiles==0.7.0
amqp==5.0.6
asgiref==3.4.0
attrs==21.2.0
autopep8==1.6.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dataclasses==0.6
fastapi==0.65.2
h11==0.12.0
idna==3.3
importlib-metadata==4.6.0
iniconfig==1.1.1
itsdangerous==2.0.1
joblib==1.0.1
kombu==5.1.0
llvmlite==0.36.0
numba==0.53.1
numpy==1.19.5
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.20
py==1.10.0
pycodestyle==2.8.0
pydantic==1.8.2
pynndescent==0.5.2
pyocclient==0.4
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.1
pytz==2021.3
redis==3.5.3
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
starlette==0.14.2
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.10.0.0
umap-learn==0.5.1
urllib3==1.26.7
uvicorn==0.14.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
